outsource components from upload page
